에러 있음
- WindowUpdate()에서 간헐적으로 죽음.
- ifstream에서 open()이 될 때도 있고 안 될 때도 있고 그럼. 대환장파티

+++ Morphosis/SourceFile/Animation/AnimationController.cpp
@@ -58,7 +58,7 @@ XMMATRIX CAnimationController::GetOffset(int boneIdx)
 }
 float CAnimationController::GetEndTime()
 {
-	return m_AnimData[m_AnimState]->m_KeyTime[m_AnimData[m_AnimState]->m_nKeyTime - 1];
+	return static_cast<float>(m_AnimData[m_AnimState]->m_KeyTime[m_AnimData[m_AnimState]->m_nKeyTime - 1]);
 }
 int CAnimationController::GetPrevIdx(float time)
 {
